start server or client options when program start
